Adding new protected async method

+++ src/resolvers/Mutation.js
@@ -90,7 +90,19 @@ const Mutation = {
             }
         }, info)
     },
-    deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma, request }, info) {
+       const userId = getUserId(request)
+       const postExists = await prisma.exists.Post({
+          id: args.id,
+          author: {
+              id: userId
+          }
+       })
+
+       if (!postExists) {
+           throw new Error('Unable to delete post')
+       }
+
        return prisma.mutation.deletePost({
           where: {
               id: args.id
